move spynnaker_plotting
