+++ p8_integration_tests/quick_test/test_param_vertex_retrieval_from_board_after_1_run/test_synfire_20n_20pc_delays_no_delay_extension_all_recording.py
@@ -23,7 +23,7 @@ import matplotlib.pyplot as plt
 import numpy
 from pyNN.utility.plotting import Figure
 import spynnaker.spike_checker as spike_checker
-from spynnaker8.spynnaker_plotting import SpynnakerPanel
+from spynnaker.spynnaker_plotting import SpynnakerPanel
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.synfire_run import SynfireRunner
