+++ spynnaker/spynnaker_plotting.py
